Attempt to split up JavaScript files and import shared functions.
