+++ src/main/webapp/project.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>Project</title>
     <link rel="stylesheet" href="style.css">
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
   </head>
   <body onload="getProjectInfo()">
     <div id="navigation">
