+++ src/main/webapp/script.js
@@ -137,39 +137,26 @@ function getProjectInfo() {
  * default values will be shown.
  */
 function getTaskInfo() {
-  // Can get taskID from param, or maybe just get it from the button that was
-  // pressed to arrive on task page.
   const params = new URLSearchParams(location.search);
   const taskID = params.get('taskID');
-  // const tasks = JSON.parse(gJSONtasks);
-  // Create variables to create the URL
-  console.log('/task?taskID=' + taskID);
-  fetch('/task?taskID=' + taskID)
-      .then((response) => response.json())
-      .then((response) => {
-        const task = response.task[0];
-        // const project = response.project[0];
-        // const subtasks = response.subtask;
-        // const users = response.users;
-        // const comments = response.comments;
-        // Fill up task page
-        const title = document.getElementById('task-title-container');
-        title.innerHTML = '<h1>' + task.name + '</h1>';
-        const description =
-          document.getElementById('task-description-container');
-        description.innerText = task.description;
-        const status = document.getElementById('task-status-container');
-        status.innerText = 'Status: ' + task.status;
-        const toProject = document.getElementById('task-project-container');
-        // toProject.appendChild(getProjectReturn(project));
-        const subtaskList = document.getElementById('task-subtasks-container');
-        // subtaskList.appendChild(getTasks(subtasks));
-        const userList = document.getElementById('task-users-container');
-        // userList.appendChild(getUsers(users));
-        // const commentList =
-        //   document.getElementById('task-comments-container');
-        // commentList.appendChild(getComments(comments));
-      });
+  const tasks = JSON.parse(gJSONtasks);
+  for (task in tasks) {
+    if (tasks[task].taskID == taskID) {
+      const title = document.getElementById('task-title-container');
+      title.innerHTML = '<h1>' + tasks[task].name + '</h1>';
+      const description = document.getElementById('task-description-container');
+      description.innerText = tasks[task].description;
+      const status = document.getElementById('task-status-container');
+      status.innerText = 'Status: ' + tasks[task].status;
+      const project = document.getElementById('task-project-container');
+      project.appendChild(getProjectReturn(tasks[task]));
+      const subtasks = document.getElementById('task-subtasks-container');
+      subtasks.appendChild(getTasks(tasks[task].subtasks));
+      const users = document.getElementById('task-users-container');
+      users.appendChild(getUsers(tasks[task].users));
+      break;
+    }
+  }
 }
 
 /**
@@ -328,3 +315,5 @@ function createUserLiElement(user) {
   liElement.appendChild(buttonElement);
   return liElement;
 }
+
+export {goToHub, goToSettings, goToUser, goToProject, goToTask};
