+++ src/main/webapp/task.html
@@ -4,7 +4,8 @@
     <meta charset="UTF-8">
     <title>Task</title>
     <link rel="stylesheet" href="style.css">
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
+    <script type="module" src="script_task.js"></script>
   </head>
   <body onload="getTaskInfo()">
     <div id="navigation">
