+++ src/main/webapp/user_settings.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>User Settings</title>
     <link rel="stylesheet" href="style.css">
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
   </head>
   <body>
     <div id="navigation">
